added charts in doctor views, updated treatment record view in appointment

+++ core/models.py
@@ -310,8 +310,8 @@ class TreatmentDoctor(models.Model):
         TreatmentRecord, on_delete=models.CASCADE, related_name='doctors')
     doctor = models.ForeignKey(
         User, on_delete=models.CASCADE, related_name='treatment_assignments')
-    percent = models.FloatField(
-        validators=[MinValueValidator(0), MaxValueValidator(100)])
+    percent = models.DecimalField(decimal_places=2, max_digits=5,
+                                  validators=[MinValueValidator(0), MaxValueValidator(100)])
     amount = models.DecimalField(
         decimal_places=2, max_digits=10, blank=True, null=True)
     date_created = models.DateTimeField(
